Update patch set 4
Patch Set 4:

Once this is merged, it will fail to build until upstream patch is merged. May this be implemented in a backwards compatible way by removing the run-validations.sh script if it exist or using %exclude for this file in the %files section?

Patch-set: 4
Reviewer: Gerrit User 128 <128@270e2033-b340-4cff-9539-693957ebf0e7>